Apparently unsigned long on x64-linux is 8 byte, on windows it's 4. Still not quite right, on my 1080 TI the MD5 animation compute shader isn't outputting correct tangent space data.

+++ camera.c
@@ -142,7 +142,7 @@ int SphereBBOXIntersection(const vec3 Center, const float Radius, const vec3 BBM
 	return 0;
 }
 
-void CameraCheckCollision(Camera_t *Camera, float *Vertex, unsigned long *Face, int NumFace)
+void CameraCheckCollision(Camera_t *Camera, float *Vertex, unsigned int *Face, int NumFace)
 {
 	float distance=0.0f;
 	vec3 n;
